Fix regression with --ntasks-per-gpu
Make gres_job_ptr->ntasks_per_gres a uint16.
Change packing to uint16.
Change NO_VAL64 checks to NO_VAL16.

ntasks_per_tres was being initialized to NO_VAL16 properly, but since
it's not a NO_VAL64, the cpus_per_gres setting logic was always
triggering, causing simple "srun --gres=gpu:1" commands to fail.

Bug 7563

+++ src/common/gres.h
@@ -196,7 +196,7 @@ typedef struct gres_job_state {
 	uint64_t gres_per_socket;
 	uint64_t gres_per_task;
 	uint64_t mem_per_gres;
-	uint64_t ntasks_per_gres;
+	uint16_t ntasks_per_gres;
 
 	/*
 	 * Default GRES configuration parameters. These values are subject to
